+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # UNRELEASED
 
+-   Add support for experimental encrypted state events, introduce
+    `EncryptionSettings:.encryptStateEvents`, `RoomSettings.encryptStateEvents`,
+    and `OlmMachine.encryptStateEvent`.
+    ([#260](https://github.com/matrix-org/matrix-sdk-crypto-wasm/pull/260/files))
+
 # matrix-sdk-crypto-wasm v15.2.0
 
 -   Update matrix-rust-sdk to 0.14.0, which includes: